Use latest ENSIME (needs separate build step)
I separate scalaLibraryVersion in a variable because:

a) It is needed by newer gradle-ensime plugin (although change is not
included in this commit)
b) It is also used in the requisite of org.scala-lang:scala-compiler

In order to build the latest, fixed and neat, gradle-ensime, you must
not use (currently) the coacoas repository. Latest ensime is PR#3 in
GitHub, which is available at https://github.com/maiflai/gradle-ensime

Once you have the code, go in there and type in `gradle install`. This
will publish the code in your local maven repo (~/.m2). Then you can use
it from this project.

If you want to "fake" 0.1.8 version you must edit gradle-ensime's
gradle.properties specifying the 0.1.8 version. Elsehow, you should edit
this project's build.gradle in order to use the 0.1.9-SNAPSHOT version
which will be installed.