Use different annotations for testing than real app.
WARNING this commit does not boot the gradle bootrun task. It only
allows to run the tests; which do not run in the previous
configuration. We need to decouple both run/test configurations from
each other in order to be able to have both.

This happens because springmvc-router expects the application context to
have dispatchers registered for web, which do not get built on testing
unless a REST integration test is requested.