Use Persistent entities on FederationBean
Update TenantFilter to work together with the FederationBean

Plus an efficiency enhancement to the FederationBean.

Now a lazy val is used to compute the value of a federation just once
per request. Even that can be costly, so Redis should be in place when
under load (and used in the DAO when the time comes)